scheduler/requirements.txt
python-dotenv==0.20.0
backoff==2.1.2
psycopg2-binary==2.9.3
